pyproject.toml
[tool.poetry]
name = "liquidation-searcher"
version = "0.1.0"
description = ""
authors = ["Akagi201 <akagi201@gmail.com>"]
readme = "README.md"
packages = [{include = "liquidation_searcher", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
black = "^23.11.0"
pyyaml = "^6.0.1"
ccxt = "^4.1.55"
orderly-sdk = { git = "ssh://git@gitlab.com/orderlynetwork/orderly-tool/orderly-sdk-py.git", branch = "master" }
wootrade-sdk = "^0.1.1"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
